requirements.txt
pyautogui==0.9.54
pillow==10.1.0
tkcap==0.0.3
pydicom==2.4.3
img2pdf==0.5.0
opencv_python==4.8.1.78
matplotlib==3.8.1
pandas==2.1.2
tensorflow==2.14.0
python-xlib==0.33
